Added OpenCV C++
Wanted to die in the process but got OpenCV C++ added into the android app and running. There were two very small issues, one is where the opencv module was added in the settings file and the other was in the .mk file since the header files needed to be exported.